+++ src/components/ui/input/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/utils/cn';
+import { cn } from '@/utils';
 
 export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {}
