+++ src/utils/index.ts
@@ -1,2 +1,4 @@
 export * from './cn';
 export * from './constants';
+export * from './editor';
+export * from './getLocalDirectory';
